+++ hiv-treatment-support-system/backend/src/main/java/backend/schedule/dto/CreateScheduleRequest.java
@@ -12,6 +12,6 @@ public record CreateScheduleRequest (
 
     LocalTime slot,
     
-    int doctorId
+    long doctorId
 ){    
 }
