+++ hiv-treatment-support-system/backend/src/main/java/backend/testresult/dto/CreateTestResultRequest.java
@@ -13,6 +13,6 @@ public record CreateTestResultRequest(
 
     LocalDateTime expectedResultTime,
     
-    int scheduleId
+    long scheduleId
 ) {
 }
